Reworked code under new assumptions for simplicity
The professor during class stated that we can assume
the two machines that will be used to run the
assignment will be identical. As such, we consolidated
all of the data we send from the client to the server
into a single struct, and removed cruft that made the
code harder to understand (like sending a whence
directly as an integer instead of translating from the
client to an intermediate character just to have the
server translate that character back to the whence).